提交后台代码

+++ Code/ManageServer/src/main/java/com/huishu/ManageServer/repository/third/AttributeRepository.java
@@ -34,7 +34,7 @@ public interface AttributeRepository extends CrudRepository<AttributeEntity, Lon
 	 * @return
 	 */
 	@Query(value="select t_word_id  from  t_word_attribute  GROUP BY t_word_id ORDER BY count(t_word_id) desc limit ?1,?2",nativeQuery=true)
-	List<Long> getKeyWordId(Long number, Integer integer);
+	List<Integer> getKeyWordId(Long number, Integer integer);
 
 	/**
 	 * 获取总数id
